Typescriptify jbrowse app level component in jbrowse-web
