--- products/jbrowse-web/src/JBrowse.js
@@ -1,83 +1,0 @@
-import React, { useEffect } from 'react'
-import { getConf } from '@jbrowse/core/configuration'
-import { useQueryParam, StringParam } from 'use-query-params'
-import { App, createJBrowseTheme } from '@jbrowse/core/ui'
-import CssBaseline from '@material-ui/core/CssBaseline'
-import { ThemeProvider } from '@material-ui/core/styles'
-import { observer } from 'mobx-react'
-import { onSnapshot } from 'mobx-state-tree'
-import ShareButton from './ShareButton'
-import AdminComponent from './AdminComponent'
-
-function deleteBaseUris(config) {
-  if (typeof config === 'object') {
-    for (const key of Object.keys(config)) {
-      if (typeof config[key] === 'object') {
-        deleteBaseUris(config[key])
-      } else if (key === 'uri') {
-        delete config.baseUri
-      }
-    }
-  }
-  return config
-}
-
-const JBrowse = observer(({ pluginManager }) => {
-  const [adminKey] = useQueryParam('adminKey', StringParam)
-  const [adminServer] = useQueryParam('adminServer', StringParam)
-  const [configPath] = useQueryParam('config', StringParam)
-  const [, setSessionId] = useQueryParam('session', StringParam)
-  const { rootModel } = pluginManager
-  const { error, jbrowse, session } = rootModel || {}
-  const { id: currentSessionId } = session
-
-  useEffect(() => {
-    setSessionId(`local-${currentSessionId}`)
-    window.JBrowseRootModel = rootModel
-    window.JBrowseSession = session
-  }, [currentSessionId, rootModel, session, setSessionId])
-
-  useEffect(() => {
-    onSnapshot(jbrowse, async snapshot => {
-      if (adminKey) {
-        const response = await fetch(adminServer || `/updateConfig`, {
-          method: 'POST',
-          headers: {
-            'Content-Type': 'application/json',
-          },
-          body: JSON.stringify({
-            adminKey,
-            configPath,
-            config: deleteBaseUris(JSON.parse(JSON.stringify(snapshot))),
-          }),
-        })
-        if (!response.ok) {
-          const message = await response.text()
-          session.notify(
-            `Admin server error: ${response.status} (${response.statusText}) ${
-              message || ''
-            }`,
-          )
-        }
-      }
-    })
-  }, [jbrowse, session, adminKey, adminServer, configPath])
-
-  if (error) {
-    throw error
-  }
-
-  const theme = getConf(rootModel.jbrowse, 'theme')
-  return (
-    <ThemeProvider theme={createJBrowseTheme(theme)}>
-      <CssBaseline />
-      <App
-        session={session}
-        HeaderButtons={<ShareButton session={session} />}
-      />
-      {adminKey ? <AdminComponent pluginManager={pluginManager} /> : null}
-    </ThemeProvider>
-  )
-})
-
-export default JBrowse
